% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{run_app}
\alias{run_app}
\title{Launch the Excel2R Shiny app}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to shiny::runApp()}
}
\description{
Opens the interactive 5-step workflow in your browser.
}
\examples{
\dontrun{
excel2r::run_app()
}
}
