% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{supported_functions}
\alias{supported_functions}
\title{List supported Excel functions}
\usage{
supported_functions()
}
\value{
Character vector of function names
}
\description{
Returns the names of all Excel functions that excel2r can translate to R.
}
\examples{
excel2r::supported_functions()
}
